scyfest/requirements.txt
asgiref==3.5.0
autocrop==1.3.0
certifi==2021.10.8
defusedxml==0.7.1
Django==4.0.4
fpdf2==2.5.2
numpy==1.22.3
opencv-python-headless==4.5.5.64
pdfrw==0.4
Pillow==9.1.0
qrcode==7.3.1
sqlparse==0.4.2
